Move a bunch of types to core
