+++ src/renderer/BaseRenderLayer.ts
@@ -4,12 +4,12 @@
  */
 
 import { IRenderLayer, IColorSet, IRenderDimensions } from './Types';
-import { ITerminal, ICellData } from '../Types';
+import { ITerminal } from '../Types';
+import { ICellData } from '../core/Types';
 import { DIM_OPACITY, INVERTED_DEFAULT_COLOR, IGlyphIdentifier, DEFAULT_COLOR } from './atlas/Types';
 import BaseCharAtlas from './atlas/BaseCharAtlas';
 import { acquireCharAtlas } from './atlas/CharAtlasCache';
-import { CellData, AttributeData } from '../BufferLine';
-import { WHITESPACE_CELL_CHAR, WHITESPACE_CELL_CODE } from '../Buffer';
+import { CellData, AttributeData, WHITESPACE_CELL_CHAR, WHITESPACE_CELL_CODE } from '../BufferLine';
 import { JoinedCellData } from './CharacterJoinerRegistry';
 
 export abstract class BaseRenderLayer implements IRenderLayer {
