+++ src/renderer/CharacterJoinerRegistry.test.ts
@@ -1,3 +1,8 @@
+/**
+ * Copyright (c) 2018 The xterm.js authors. All rights reserved.
+ * @license MIT
+ */
+
 import { assert } from 'chai';
 
 import { MockTerminal, MockBuffer } from '../TestUtils.test';
@@ -6,7 +11,7 @@ import { CircularList } from '../common/CircularList';
 import { ICharacterJoinerRegistry } from './Types';
 import { CharacterJoinerRegistry } from './CharacterJoinerRegistry';
 import { BufferLine, CellData } from '../BufferLine';
-import { IBufferLine } from '../Types';
+import { IBufferLine } from '../core/Types';
 
 describe('CharacterJoinerRegistry', () => {
   let registry: ICharacterJoinerRegistry;
