+++ src/renderer/CursorRenderLayer.ts
@@ -5,7 +5,8 @@
 
 import { IColorSet, IRenderDimensions } from './Types';
 import { BaseRenderLayer } from './BaseRenderLayer';
-import { ITerminal, ICellData } from '../Types';
+import { ITerminal } from '../Types';
+import { ICellData } from '../core/Types';
 import { CellData } from '../BufferLine';
 
 interface ICursorState {
