+++ src/renderer/TextRenderLayer.ts
@@ -3,12 +3,12 @@
  * @license MIT
  */
 
-import { NULL_CELL_CODE } from '../Buffer';
 import { IColorSet, IRenderDimensions, ICharacterJoinerRegistry } from './Types';
-import { CharData, ITerminal, ICellData } from '../Types';
+import { ITerminal } from '../Types';
+import { CharData, ICellData } from '../core/Types';
 import { GridCache } from './GridCache';
 import { BaseRenderLayer } from './BaseRenderLayer';
-import { CellData, AttributeData, Content } from '../BufferLine';
+import { CellData, AttributeData, Content, NULL_CELL_CODE } from '../BufferLine';
 import { JoinedCellData } from './CharacterJoinerRegistry';
 
 /**
