+++ src/renderer/dom/DomRendererRowFactory.ts
@@ -3,10 +3,10 @@
  * @license MIT
  */
 
-import {  NULL_CELL_CODE, WHITESPACE_CELL_CHAR } from '../../Buffer';
-import { IBufferLine, ITerminalOptions } from '../../Types';
+import { ITerminalOptions } from '../../Types';
+import { IBufferLine } from '../../core/Types';
 import { INVERTED_DEFAULT_COLOR } from '../atlas/Types';
-import { CellData, AttributeData } from '../../BufferLine';
+import { CellData, AttributeData, NULL_CELL_CODE, WHITESPACE_CELL_CHAR } from '../../BufferLine';
 
 export const BOLD_CLASS = 'xterm-bold';
 export const DIM_CLASS = 'xterm-dim';
